refactor store and related functions
