+++ package.json
@@ -41,7 +41,7 @@
     "rollup-plugin-postcss": "^3.1.8",
     "rollup-plugin-terser": "^7.0.1",
     "tslib": "^2.0.1",
-    "typescript": "^4.0.3"
+    "typescript": "^4.1.0-dev.20201005"
   },
   "dependencies": {
     "nodegit": "^0.27.0"
