+++ src/Data/Renderer/store.ts
@@ -1,7 +1,5 @@
-import { IpcAction, BranchesObj, BranchObj, IpcActionReturn, PatchObj, IpcActionReturnError, Locks, RepoStatus } from "../Actions";
-import { registerHandler, sendAsyncMessage, attach } from ".";
-import { ipcRenderer } from "electron";
-import { WindowEvents, WindowArguments } from "../WindowEventTypes";
+import { IpcAction, BranchesObj, BranchObj, PatchObj, Locks, RepoStatus } from "../Actions";
+import { sendAsyncMessage } from "./IPC";
 
 export type DialogWindow = {
     title: string
@@ -104,11 +102,6 @@ export function openRepo(repoPath: string) {
     setLock(Locks.MAIN);
     sendAsyncMessage(IpcAction.OPEN_REPO, repoPath);
 }
-export function loadBranches() {
-    console.log("fetched");
-    setLock(Locks.MAIN);
-    sendAsyncMessage(IpcAction.LOAD_BRANCHES);
-}
 
 export function checkoutBranch(branch: string) {
     setLock(Locks.MAIN);
@@ -152,74 +145,6 @@ export function continueRebase() {
     sendAsyncMessage(IpcAction.CONTINUE_REBASE);
 }
 
-function setStatus(status: RepoStatus) {
-    const repo = store.repo;
-    if (repo) {
-        repo.status = status;
-        setState({
-            repo
-        });
-    }
-}
-
-function loadHunks(data: IpcActionReturn[IpcAction.LOAD_HUNKS]) {
-    if (store.currentFile && data.hunks) {
-        store.currentFile.patch.hunks = data.hunks;
-        setState({
-            currentFile: store.currentFile
-        });
-    }
-}
-
-function repoOpened(result: IpcActionReturn[IpcAction.OPEN_REPO] | IpcActionReturnError) {
-    clearLock(Locks.MAIN);
-    if ("error" in result) {
-        console.warn(result);
-    } else if (result.opened) {
-        localStorage.setItem("recent-repo", result.path);
-        setState({
-            repo: {
-                path: result.path,
-                status: result.status
-            }
-        });
-        loadBranches();
-        refreshWorkdir();
-    } else {
-        setState({
-            repo: null
-        });
-    }
-}
-function mapHeads(heads: any, refs: BranchObj[]) {
-    for (const ref of refs) {
-        if (!heads[ref.headSHA]) {
-            heads[ref.headSHA] = [];
-        }
-        heads[ref.headSHA].push(ref);
-    }
-}
-function branchesLoaded(branches: BranchesObj) {
-    clearLock(Locks.MAIN);
-    console.log("loaded branches");
-    const heads:any = {};
-    mapHeads(heads, branches.local);
-    mapHeads(heads, branches.remote);
-    mapHeads(heads, branches.tags);
-    setState({
-        branches,
-        heads
-    });
-}
-function updateCurrentBranch(result: IpcActionReturn[IpcAction.CHECKOUT_BRANCH]) {
-    clearLock(Locks.MAIN);
-    if (result && store.branches) {
-        store.branches.head = result;
-        setState({
-            branches: store.branches
-        });
-    }
-}
 export function setLock(lock: keyof StoreType["locks"], event?:any) {
     const locks = store.locks;
     locks[lock]++;
@@ -236,16 +161,7 @@ export function clearLock(lock: keyof StoreType["locks"]) {
         locks,
     });
 }
-export function refreshWorkdir() {
-    sendAsyncMessage(IpcAction.REFRESH_WORKDIR);
-}
-export function openSettings() {
-    console.log("open settings");
-}
-export function pullHead() {
-    setLock(Locks.MAIN);
-    sendAsyncMessage(IpcAction.PULL);
-}
+
 export function createBranch(fromSha: string, name: string) {
     sendAsyncMessage(IpcAction.CREATE_BRANCH, {
         sha: fromSha,
@@ -264,38 +180,6 @@ export function deleteBranch(name: string) {
     });
 }
 
-function openDialogCompareRevisions() {
-    openDialogWindow({
-        title: "Compare revisions:",
-        confirmCb: (data: any) => {
-            if (data.branchName)
-            {
-                const [from,to] = data.branchName.split("..");
-                if (from && to) {
-                    sendAsyncMessage(IpcAction.OPEN_COMPARE_REVISIONS, {
-                        from,
-                        to
-                    });
-                }
-            }
-            closeDialogWindow();
-        },
-        cancelCb: () => {
-            closeDialogWindow();
-        }
-    });
-}
-function handleCompareRevisions(data: any) {
-    if ("error" in data) {
-        console.warn(data.error);
-    } else {
-        setState({
-            comparePatches: data,
-            selectedBranch: undefined,
-        });
-    }
-}
-
 export function openDialogWindow(dialogWindow: StoreType["dialogWindow"]) {
     setState({
         dialogWindow
@@ -306,38 +190,3 @@ export function closeDialogWindow() {
         dialogWindow: null
     });
 }
-
-function addWindowEventListener<T extends WindowEvents>(event: T, cb: (args: WindowArguments[T]) => void) {
-    // @ts-ignore
-    ipcRenderer.on(event, (_, args) => cb(args, event));
-}
-
-attach();
-addWindowEventListener("repo-opened", repoOpened);
-addWindowEventListener("repo-fetch-all", loadBranches);
-addWindowEventListener("refresh-workdir", refreshWorkdir);
-addWindowEventListener("open-settings", openSettings);
-addWindowEventListener("app-lock-ui", setLock);
-addWindowEventListener("app-unlock-ui", clearLock);
-addWindowEventListener("pull-head", pullHead);
-addWindowEventListener("begin-compare-revisions", openDialogCompareRevisions);
-addWindowEventListener("fetch-status", (stats: WindowArguments["fetch-status"]) => {
-    if (stats.receivedObjects == stats.totalObjects) {
-        console.log(`Resolving deltas ${stats.indexedDeltas}/${stats.totalDeltas}`);
-    } else if (stats.totalObjects > 0) {
-        console.log(`Received ${stats.receivedObjects}/${stats.totalObjects} objects (${stats.indexedObjects}) in ${stats.receivedBytes} bytes`);
-    }
-});
-
-registerHandler(IpcAction.OPEN_REPO, repoOpened);
-registerHandler(IpcAction.LOAD_BRANCHES, branchesLoaded);
-registerHandler(IpcAction.CHECKOUT_BRANCH, updateCurrentBranch);
-registerHandler(IpcAction.LOAD_HUNKS, loadHunks);
-registerHandler(IpcAction.PULL, loadBranches);
-registerHandler(IpcAction.PUSH, loadBranches);
-registerHandler(IpcAction.CREATE_BRANCH, loadBranches);
-registerHandler(IpcAction.CREATE_BRANCH_FROM_REF, loadBranches);
-registerHandler(IpcAction.DELETE_REF, loadBranches);
-registerHandler(IpcAction.ABORT_REBASE, setStatus);
-registerHandler(IpcAction.CONTINUE_REBASE, setStatus);
-registerHandler(IpcAction.OPEN_COMPARE_REVISIONS, handleCompareRevisions);
